[native][rust] create expect implementation which actually logs on android
previously we just unwrapped everything because we wouldn't get the error messages anyways. this implementation does depend on jni calls into java
since at least for now i know of no ergonomic way to use ndk logging functions from rust, this isn't a massive issue as the overhead is minor and it
should be save to assume that the Log api is stable.